requirements.txt
boto3==1.17.105
botocore==1.20.105
jmespath==0.10.0
Pillow==8.3.0
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.6
